requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9
pandas==2.2.2
openpyxl==3.1.5

redis==5.0.8
celery==5.4.0
requests

aiofiles
reportlab
weasyprint
google-api-python-client
google-auth
fonttools 
brotli

python-multipart==0.0.9
httpx==0.27.2

loguru==0.7.2
email-validator==2.2.0
Jinja2>=3.1
python-jose[cryptography]==3.3.0
pymupdf==1.24.10

rembg==2.0.57

onnxruntime>=1.17.0
Pillow==10.2.0
